# openapi.yaml
openapi: 3.0.3
info:
  title: ACS Multi-Vendor TR-069/CWMP REST API
  version: 1.0.0
  description: |
    REST API internal (Echo v5) untuk Auto Configuration Server (ACS) TR-069/CWMP
    multi-vendor (ZTE, Huawei, FiberHome, Nokia, C-Data, dst). Dikonsumsi oleh
    sistem BSS/OSS, portal NOC, dan integrasi internal lain — **BUKAN** endpoint
    yang dipakai CPE (ONT/ONU/router) itu sendiri.

    **PENTING — beda dengan endpoint CWMP:** perangkat CPE berkomunikasi dengan
    ACS lewat endpoint `/cwmp` (SOAP/XML, protokol TR-069/CWMP standar Broadband
    Forum, sesi stateful) yang **TIDAK** didokumentasikan di file ini sama
    sekali — beda protokol, beda tujuan, beda konsumen. Dokumen ini HANYA
    mencakup `/api/v1/*` (REST + JSON, kecuali `GET /api/v1/metrics` yang
    formatnya teks Prometheus).

    ## Autentikasi

    Hampir seluruh endpoint mewajibkan header `Authorization: Bearer <token>`,
    berupa salah satu dari:
    - **JWT sesi user** — didapat dari `POST /auth/login`, berlaku sesuai masa
      berlaku yang dikonfigurasi (`ACS_JWT_EXPIRY`, default 8 jam di `.env.example`).
    - **API token** — string berawalan `acs_...`, diterbitkan lewat
      `POST /auth/tokens` (ADMIN/SUPERADMIN), tidak pernah kedaluwarsa kecuali
      `expires_at` diisi.

    Empat endpoint publik tanpa autentikasi, semuanya READ-ONLY atau titik masuk
    login (tidak ada endpoint MUTASI tanpa autentikasi): `POST /auth/login`
...
    aman).

...

    ## RBAC

...

...

...
    skema `Task`.

    ## Server

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /provisioning-profiles:
    get:
      operationId: listProfiles
      tags: [Provisioning]
      summary: Daftar provisioning profile
      description: |
        Semua role terautentikasi. Menampilkan profile milik tenant sendiri
        DITAMBAH profile global (`tenant_id: null`, dibuat superadmin sebagai
        default lintas tenant). Parameter `tenant_id` di query hanya efektif
        untuk superadmin.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/tenantIdQuery'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: Daftar profile.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProvisioningProfileListResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: "Actor non-superadmin tanpa `tenant_id`."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    post:
      operationId: createProfile
      tags: [Provisioning]
      summary: Membuat provisioning profile baru
      description: |
        ADMIN/SUPERADMIN. Non-superadmin hanya boleh membuat profile untuk
        tenant miliknya sendiri (`tenant_id: null`/profile global HANYA bisa
        dibuat superadmin).
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProfileRequest' }
      responses:
        '201':
          description: Profile dibuat.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProvisioningProfile' }
        '400':
          description: "Payload tidak valid atau `name` kosong."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN, atau non-superadmin mencoba membuat profile untuk tenant lain/profile global.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /provisioning-profiles/{id}:
    get:
      operationId: getProfile
      tags: [Provisioning]
      summary: Detail profile beserta daftar parameternya
      description: |
        Semua role terautentikasi — boleh membaca profile tenant sendiri ATAU
        profile global (`tenant_id: null`).
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/profileIdPath'
      responses:
        '200':
          description: Detail profile + parameter.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProvisioningProfileDetail' }
        '400':
          description: "`id` tidak valid."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Profile milik tenant lain (bukan global).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Profile tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    put:
      operationId: updateProfile
      tags: [Provisioning]
      summary: Update profile (full update) beserta parameternya
      description: |
        ADMIN/SUPERADMIN. Profile global (`tenant_id: null`) HANYA bisa diedit
        superadmin (mencegah satu tenant diam-diam mengubah default bersama).
        Bila field `parameters` dikirim (non-null), SELURUH parameter lama
        profile diganti (replace, bukan merge); bila tidak dikirim, parameter
        yang sudah ada tidak disentuh.

        **PENTING (FR-18):** mengubah profile ini TIDAK otomatis mendorong
        ulang konfigurasi ke device yang SUDAH terprovisioning dengan profile
        ini sebelumnya — ini keputusan produk yang disengaja. Untuk menerapkan
        perubahan ke device tertentu, panggil eksplisit
        `POST /devices/:id/apply-profile/:profileId`.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/profileIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfileRequest' }
      responses:
        '204':
          description: Profile berhasil diupdate.
        '400':
          description: "`id` tidak valid atau payload tidak valid."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN, profile tenant lain, atau non-superadmin mencoba mengedit profile global.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Profile tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      operationId: deleteProfile
      tags: [Provisioning]
      summary: Soft-delete profile
      description: ADMIN/SUPERADMIN, scope sama seperti `PUT /provisioning-profiles/:id`.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/profileIdPath'
      responses:
        '204':
          description: Profile berhasil di-soft-delete.
        '400':
          description: "`id` tidak valid."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN, profile tenant lain, atau non-superadmin mencoba menghapus profile global.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Profile tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /devices/{id}/apply-profile/{profileId}:
    post:
      operationId: applyProfile
      tags: [Provisioning]
      summary: Menerapkan provisioning profile ke device (eksplisit)
      description: |
        ADMIN/SUPERADMIN. Mengantre SATU task `SET_PARAMETER_VALUES` berisi
        seluruh parameter profile, lalu mengeset `provisioning_profile_id`
        device ke profile ini. SELALU dipanggil eksplisit (manual, atau
        otomatis dari evaluasi Zero-Touch Provisioning saat BOOTSTRAP) — lihat
        catatan FR-18 di `PUT /provisioning-profiles/:id`. Ditolak `429` bila
        kuota task queue tenant sudah tercapai.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/deviceIdPath'
        - name: profileId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '201':
          description: Task SET_PARAMETER_VALUES yang diantre.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '400':
          description: "`id` device/`profileId` tidak valid, atau profile tidak punya parameter untuk diterapkan."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN, device/profile milik tenant lain.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Device atau profile tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '429':
          description: Kuota task queue tenant sudah tercapai.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ============================== ZERO-TOUCH ==============================
  /zero-touch-rules:
    get:
      operationId: listZTRules
      tags: [ZeroTouch]
      summary: Daftar aturan Zero-Touch Provisioning
      description: |
        Semua role terautentikasi. Menampilkan rule tenant sendiri + rule
        global (`tenant_id: null`), terurut `priority` ASC. `tenant_id` di
        query hanya efektif untuk superadmin.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/tenantIdQuery'
      responses:
        '200':
          description: Daftar rule (array polos, TIDAK dibungkus `data`/`total`).
          content:
            application/json:
              schema:
                type: array
                nullable: true
                items: { $ref: '#/components/schemas/ZeroTouchRule' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: "Actor non-superadmin tanpa `tenant_id`."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    post:
      operationId: createZTRule
      tags: [ZeroTouch]
      summary: Membuat aturan Zero-Touch Provisioning baru
      description: |
        ADMIN/SUPERADMIN. Rule dievaluasi urut `priority` ASC (angka lebih
        kecil dievaluasi lebih dulu) saat device baru mengirim event
        `0 BOOTSTRAP`. Non-superadmin hanya boleh membuat rule untuk tenant
        miliknya sendiri.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ZeroTouchRuleRequest' }
      responses:
        '201':
          description: Rule dibuat.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ZeroTouchRule' }
        '400':
          description: "Payload tidak valid, `trigger_event_id` kosong/0, atau salah satu field pola melebihi batas panjang kolom. Catatan: `provisioning_profile_id` OPSIONAL sejak migrations/0009 (rule boleh hanya memicu reboot/firmware)."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN, atau non-superadmin mencoba membuat rule untuk tenant lain/global.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /zero-touch-rules/{id}:
    put:
      operationId: updateZTRule
      tags: [ZeroTouch]
      summary: Update aturan Zero-Touch Provisioning (full update)
      description: ADMIN/SUPERADMIN, scope tenant sama seperti `POST /zero-touch-rules`.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ztRuleIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ZeroTouchRuleRequest' }
      responses:
        '204':
          description: Rule berhasil diupdate.
        '400':
          description: "`id` tidak valid atau payload tidak valid."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN, atau rule milik tenant lain.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Rule tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      operationId: deleteZTRule
      tags: [ZeroTouch]
      summary: Soft-delete aturan Zero-Touch Provisioning
      description: ADMIN/SUPERADMIN, scope sama seperti `PUT /zero-touch-rules/:id`.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ztRuleIdPath'
      responses:
        '204':
          description: Rule berhasil di-soft-delete.
        '400':
          description: "`id` tidak valid."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN, atau rule milik tenant lain.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Rule tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ============================== CATALOG ==============================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT sesi user (dari `POST /auth/login`) ATAU API token statis
        berawalan `acs_...` (dari `POST /auth/tokens`) — keduanya dikirim
        dengan cara yang sama lewat header `Authorization: Bearer <token>`.
  parameters:
    page:
      name: page
      in: query
      description: Halaman (1-based). Default 1.
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: page_size
      in: query
      description: Ukuran halaman. Default 50, maksimum 200 (nilai di luar rentang ini jatuh ke default 50).
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    tenantIdQuery:
      name: tenant_id
      in: query
      description: |
        Filter tenant. HANYA efektif untuk actor `SUPERADMIN` — actor role lain
        selalu dipaksa ke tenant miliknya sendiri, nilai ini diabaikan.
      schema: { type: integer, format: int64 }
...
    deviceIdPath:
      name: id
      in: path
      required: true
      description: ID numerik internal device.
      schema: { type: integer, format: int64 }
...
    profileIdPath:
      name: id
      in: path
      required: true
      description: ID numerik internal provisioning profile.
      schema: { type: integer, format: int64 }
    ztRuleIdPath:
      name: id
      in: path
      required: true
      description: ID numerik internal zero-touch rule.
      schema: { type: integer, format: int64 }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: |
        Bentuk error SERAGAM untuk seluruh endpoint (`echo.NewHTTPError`) —
        selalu `{"message": "..."}`, tidak ada kode error mesin terpisah dari
        `message` teks bebas saat ini (perhatikan: teks pesan sebagian besar
        Bahasa Indonesia, konsisten dengan gaya pesan aplikasi ini).
      properties:
        message:
          type: string
          example: "domain: data tidak ditemukan"
      required: [message]
...
    Task:
      type: object
      description: |
        Satu item antrean RPC CWMP. `parameters` dan `response` bertipe
        `domain.JSONRawMessage` — di-emit sebagai **objek JSON apa adanya**
        (bukan string ter-base64-kan; asimetri base64 lama sudah diperbaiki,
        lihat log perubahan `ROADMAP.md` 2026-08-23). `response` bernilai
        `null` selama task belum selesai. Untuk role `NOC`/`VIEWER`, key
        `url`/`username`/`password` di dalam `parameters` di-redact jadi
        `"[redacted]"`.
      properties:
        id: { type: integer, format: int64 }
        task_uuid: { type: string, format: uuid }
        device_id: { type: integer, format: int64 }
        task_type_id: { type: integer, format: int64 }
        task_status_id: { type: integer, format: int64 }
        priority: { type: integer, minimum: 1, maximum: 9, description: "1 = tertinggi, 9 = terendah." }
        parameters:
          type: object
          additionalProperties: true
          nullable: true
          description: "Objek JSON parameter RPC (mis. daftar nama parameter, key-value SetParameterValues)."
        response:
          type: object
          additionalProperties: true
          nullable: true
          description: "Objek JSON respons RPC dari CPE. `null` selama task belum selesai."
        error_message: { type: string, nullable: true }
        retry_count: { type: integer }
        max_retries: { type: integer }
        scheduled_at: { type: string, format: date-time, nullable: true }
        expires_at: { type: string, format: date-time, nullable: true }
        sent_at: { type: string, format: date-time, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
      allOf:
        - $ref: '#/components/schemas/Audit'
      required: [id, task_uuid, device_id, task_type_id, task_status_id, priority]
...
    ProvisioningProfile:
      type: object
      properties:
        id: { type: integer, format: int64 }
        profile_uuid: { type: string, format: uuid }
        tenant_id: { type: integer, format: int64, nullable: true, description: "NULL berarti profile global lintas tenant (hanya bisa dibuat/diedit superadmin)." }
        vendor_id: { type: integer, format: int64, nullable: true }
        device_model_id: { type: integer, format: int64, nullable: true }
        name: { type: string }
        description: { type: string, nullable: true }
        is_default: { type: boolean }
        is_active: { type: boolean }
      allOf:
        - $ref: '#/components/schemas/Audit'
      required: [id, profile_uuid, name, is_default, is_active]
    ProvisioningProfileListResponse:
      type: object
      properties:
        data:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/ProvisioningProfile' }
        total: { type: integer }
      required: [data, total]
...
    ProvisioningProfileDetail:
      type: object
      description: "Response gabungan khusus `GET /provisioning-profiles/:id` — BUKAN `ProvisioningProfile` polos."
      properties:
        profile: { $ref: '#/components/schemas/ProvisioningProfile' }
        parameters:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/ProvisioningProfileParameter' }
      required: [profile, parameters]
    ProfileParameterInput:
      type: object
      properties:
        parameter_name: { type: string }
        parameter_value: { type: string, nullable: true }
        apply_order: { type: integer, default: 0 }
      required: [parameter_name]
    CreateProfileRequest:
      type: object
      properties:
        tenant_id:
          type: integer
          format: int64
          nullable: true
          description: "NULL = profile global (superadmin only). Non-superadmin WAJIB isi dengan tenant miliknya sendiri."
        vendor_id: { type: integer, format: int64, nullable: true }
        device_model_id: { type: integer, format: int64, nullable: true }
        name: { type: string, maxLength: 128 }
        description: { type: string, nullable: true, maxLength: 255 }
        is_default: { type: boolean, default: false }
        parameters:
          type: array
          items: { $ref: '#/components/schemas/ProfileParameterInput' }
      required: [name]
    UpdateProfileRequest:
      type: object
      description: "FULL update untuk field skalar (bukan partial — kirim seluruh state). `parameters` NULL/tidak dikirim = parameter lama TIDAK disentuh; array (termasuk kosong `[]`) = REPLACE seluruh parameter lama."
      properties:
        vendor_id: { type: integer, format: int64, nullable: true }
        device_model_id: { type: integer, format: int64, nullable: true }
        name: { type: string, maxLength: 128 }
        description: { type: string, nullable: true, maxLength: 255 }
        is_default: { type: boolean }
        is_active: { type: boolean }
        parameters:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/ProfileParameterInput' }
    ZeroTouchRule:
      type: object
      properties:
        id: { type: integer, format: int64 }
        tenant_id: { type: integer, format: int64, nullable: true, description: "NULL berarti rule global lintas tenant." }
        vendor_id: { type: integer, format: int64, nullable: true }
        device_model_id: { type: integer, format: int64, nullable: true }
        oui: { type: string, nullable: true }
        serial_pattern: { type: string, nullable: true, description: "Pola SQL LIKE (`%`/`_`), dicocokkan case-insensitive." }
        software_version_pattern: { type: string, nullable: true, description: "Pola SQL LIKE thd devices.software_version (migrations/0009)." }
        match_parameter_name: { type: string, nullable: true, description: "Precondition opsional — device harus punya device_parameters dgn nama ini yang nilainya cocok match_parameter_value_pattern. Wajib berpasangan." }
        match_parameter_value_pattern: { type: string, nullable: true, description: "Pola SQL LIKE thd nilai parameter match_parameter_name. Wajib berpasangan." }
        provisioning_profile_id: { type: integer, format: int64, nullable: true, description: "NULLABLE sejak migrations/0009 — rule boleh hanya memicu reboot dan/atau firmware push." }
        post_apply_reboot: { type: boolean, description: "Jika true, enqueue task REBOOT setelah aksi lain diterapkan." }
        firmware_file_id: { type: integer, format: int64, nullable: true, description: "Jika diisi, enqueue task Download firmware ini ke device yang cocok." }
        trigger_event_id: { type: integer, format: int64, description: "FK ref_ztp_trigger_event — kapan rule dievaluasi (BOOTSTRAP_ONLY / BOOTSTRAP_OR_BOOT / EVERY_INFORM)." }
        priority: { type: integer, description: "Angka lebih kecil dievaluasi lebih dulu." }
        is_active: { type: boolean }
      allOf:
        - $ref: '#/components/schemas/Audit'
      required: [id, trigger_event_id, priority, is_active]
...